data_engineering_4/requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
boto3==1.29.7
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
